@@ -19,7 +19,7 @@ class DepartamentosType extends AbstractType
 	 */
 	public function __construct(DepartamentosRepository $departamentosRepository) {
 		$result = $departamentosRepository->findAll();
-		$departamentosSelect = array('- Selecciona -' => '');
+		$departamentosSelect = ['- Selecciona -' => ''];
 		
 		foreach($result as $departamento) {
 			$departamentosSelect[$departamento->getNombreDepartamento()] = $departamento->getNombreDepartamento();
@@ -29,7 +29,7 @@ class DepartamentosType extends AbstractType
 		
 	}
 	
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('nombreDepartamento', ChoiceType::class, [
@@ -39,7 +39,7 @@ class DepartamentosType extends AbstractType
         ;
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => Departamentos::class,
