@@ -26,7 +26,7 @@ class UserType extends AbstractType
 	 */
 	public function __construct(DepartamentosRepository $departamentosRepository) {
 		$result = $departamentosRepository->findAll();
-		$departamentosSelect = array('- Selecciona -' => '');
+		$departamentosSelect = ['- Selecciona -' => ''];
 		
 		foreach($result as $departamento) {
 			$departamentosSelect[$departamento->getNombreDepartamento()] = $departamento->getNombreDepartamento();
@@ -36,7 +36,7 @@ class UserType extends AbstractType
 		
 	}
 	
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder            
             ->add('nombre',TextType::class)
@@ -64,7 +64,7 @@ class UserType extends AbstractType
         ;
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => User::class,
