@@ -23,10 +23,15 @@ use Symfony\Component\Security\Core\User\UserInterface;
 class AdminDashboardController extends AbstractDashboardController
 {
 	private $security;
+ /**
+  * @var \EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator
+  */
+ private $adminUrlGenerator;
 
-	public function __construct(Security $security)
+	public function __construct(Security $security, \EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator $adminUrlGenerator)
 	{       
 	  $this->security = $security;
+   $this->adminUrlGenerator = $adminUrlGenerator;
 	}	
 	
     /**
@@ -39,14 +44,14 @@ class AdminDashboardController extends AbstractDashboardController
 			$this->addFlash('warning', 'Debes ser administrador para acceder.');
 			return $this->redirectToRoute('main_menu');
 		}
-		
+
         // redirect to some CRUD controller
-        $routeBuilder = $this->get(AdminUrlGenerator::class);
+        $this->adminUrlGenerator;
         //return $this->redirect($routeBuilder->setController(DepartamentosCrudController::class)->generateUrl());
         return $this->render('dashboard/base.html.twig', [
         	'my_own_data' => '',
         ]);
-        
+
         // you can also redirect to different pages depending on the current user
         /*if ('jane' === $this->getUser()->getUsername()) {
             return $this->redirect('...');
@@ -79,7 +84,7 @@ class AdminDashboardController extends AbstractDashboardController
 			$this->addFlash('warning', 'Debes ser administrador para acceder.');
 			return $this->generateUrl('main_menu');
 		}
-    	
+
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');       
         yield MenuItem::linkToCrud('Departamento', 'fas fa-desktop', Departamentos::class);               
         yield MenuItem::linkToCrud('Familia', 'fas fa-list', Familia::class);
