@@ -12,7 +12,7 @@ class UserCrudController extends AbstractCrudController
     {
         return User::class;
     }
-    
+
     public function configureFields(string $pageName): iterable
     {
         return [
@@ -24,7 +24,7 @@ class UserCrudController extends AbstractCrudController
             'departamento',            
         ];
     }
-    
+
     public function configureFilters(Filters $filters): Filters
     {
         return $filters
