@@ -9,7 +9,7 @@ use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 
 class AbreviaType extends AbstractType
 {
-	public function buildForm(FormBuilderInterface $builder, array $options)
+	public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder        	
             ->add('abrev', ChoiceType::class, [
@@ -29,7 +29,7 @@ class AbreviaType extends AbstractType
 		;                        
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             //'data_class' => Cliente::class,
