@@ -11,7 +11,7 @@ use Symfony\Component\Validator\Constraints\File;
 
 class ChangeUserPasswdType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder                        
             ->add('password', PasswordType::class)
@@ -19,7 +19,7 @@ class ChangeUserPasswdType extends AbstractType
         ;
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => User::class,
